TVS/tx: extract "local" conn-provider
