+++ versioned/persist/tx/src/main/java/org/projectnessie/versioned/persist/tx/TxConnectionProvider.java
@@ -15,30 +15,20 @@
  */
 package org.projectnessie.versioned.persist.tx;
 
-import io.agroal.api.AgroalDataSource;
-import io.agroal.api.configuration.AgroalConnectionFactoryConfiguration.TransactionIsolation;
-import io.agroal.api.configuration.supplier.AgroalConnectionFactoryConfigurationSupplier;
-import io.agroal.api.configuration.supplier.AgroalConnectionPoolConfigurationSupplier;
-import io.agroal.api.configuration.supplier.AgroalDataSourceConfigurationSupplier;
-import io.agroal.api.security.NamePrincipal;
-import io.agroal.api.security.SimplePassword;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.text.MessageFormat;
-import java.time.Duration;
-import java.time.temporal.ChronoUnit;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import javax.sql.DataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class TxConnectionProvider implements AutoCloseable {
+public abstract class TxConnectionProvider implements AutoCloseable {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(TxConnectionProvider.class);
 
@@ -105,60 +95,8 @@ public class TxConnectionProvider implements AutoCloseable {
 
   /**
    * Borrow a connection from the {@link TxConnectionProvider} implementation.
+   *
    * @return borrowed {@link Connection}
    */
-  public Connection borrowConnection() throws SQLException {
-    throw new UnsupportedOperationException("Must be implemented.");
-  }
-
-  public static class LocalConnectionProvider extends TxConnectionProvider {
-    private DataSource dataSource;
-
-    @Override
-    public void configure(TxDatabaseAdapterConfig config) {
-      AgroalDataSourceConfigurationSupplier dataSourceConfiguration =
-          new AgroalDataSourceConfigurationSupplier();
-      AgroalConnectionPoolConfigurationSupplier poolConfiguration =
-          dataSourceConfiguration.connectionPoolConfiguration();
-      AgroalConnectionFactoryConfigurationSupplier connectionFactoryConfiguration =
-          poolConfiguration.connectionFactoryConfiguration();
-
-      // configure pool
-      poolConfiguration
-          .initialSize(config.getPoolInitialSize())
-          .maxSize(config.getPoolMaxSize())
-          .minSize(config.getPoolMinSize())
-          .maxLifetime(Duration.of(config.getPoolConnectionLifetimeMinutes(), ChronoUnit.MINUTES))
-          .acquisitionTimeout(
-              Duration.of(config.getPoolAcquisitionTimeoutSeconds(), ChronoUnit.SECONDS));
-
-      // configure supplier
-      connectionFactoryConfiguration.jdbcUrl(config.getJdbcUrl());
-      if (config.getJdbcUser() != null) {
-        connectionFactoryConfiguration.credential(new NamePrincipal(config.getJdbcUser()));
-        connectionFactoryConfiguration.credential(new SimplePassword(config.getJdbcPass()));
-      }
-      connectionFactoryConfiguration.jdbcTransactionIsolation(
-          TransactionIsolation.valueOf(config.getPoolTransactionIsolation()));
-      connectionFactoryConfiguration.autoCommit(false);
-
-      try {
-        dataSource = AgroalDataSource.from(dataSourceConfiguration.get());
-      } catch (SQLException e) {
-        throw new RuntimeException(e);
-      }
-    }
-
-    @Override
-    public void close() throws Exception {
-      if (dataSource instanceof AutoCloseable) {
-        ((AutoCloseable) dataSource).close();
-      }
-    }
-
-    @Override
-    public Connection borrowConnection() throws SQLException {
-      return dataSource.getConnection();
-    }
-  }
+  public abstract Connection borrowConnection() throws SQLException;
 }
