+++ versioned/persist/tx/src/main/java/org/projectnessie/versioned/persist/tx/local/DefaultLocalDatabaseAdapterConfig.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2020 Dremio
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.projectnessie.versioned.persist.tx.local;
+
+import org.immutables.value.Value;
+
+/**
+ * Default "implementation" of {@link LocalDatabaseAdapterConfig}. This is a separate
+ * immutable-annotated interface, to prevent incoherences in generated types if the
+ * immutable-annotation would be present on {@link LocalDatabaseAdapterConfig}.
+ */
+@Value.Immutable(lazyhash = true)
+public interface DefaultLocalDatabaseAdapterConfig extends LocalDatabaseAdapterConfig {}
